chore: Image actions tracking
